+++ outputs.tf
@@ -1,19 +1,24 @@
 output "distribution_id" {
   description = "ID of the CloudFront distribution"
-  value       = null
+  value       = aws_cloudfront_distribution.this.id
 }
 
 output "arn" {
   description = "ARN of the CloudFront distribution"
-  value       = null
+  value       = aws_cloudfront_distribution.this.arn
 }
 
 output "domain_name" {
   description = "CloudFront-assigned domain name"
-  value       = null
+  value       = aws_cloudfront_distribution.this.domain_name
 }
 
 output "hosted_zone_id" {
   description = "Route 53 hosted zone ID of the distribution"
-  value       = null
+  value       = aws_cloudfront_distribution.this.hosted_zone_id
+}
+
+output "status" {
+  description = "Deployment status of the distribution"
+  value       = aws_cloudfront_distribution.this.status
 }